+++ sdk/issuingPurchase/issuingPurchase.js
@@ -13,11 +13,11 @@ class IssuingPurchase extends Resource {
      * @description Displays the IssuingPurchase objects created in your Workspace.
      *
      * Attributes (return-only):
-     * @param id [string]: unique id returned when IssuingPurchase is created. ex: '5656565656565656'
      * @param holderName [string]: cardholder's name. ex: 'Tony Stark'
+     * @param productId [string]: unique card product number (BIN) registered within the card network. ex: '53810200'
      * @param cardId [string]: unique id returned when IssuingCard is created. ex: '5656565656565656'
      * @param cardEnding [string]: last 4 digits of the card number. ex: '1234'
-     * @param purpose [string]: purchase purpose. ex: "purchase"
+     * @param purpose [string]: purchase purpose. ex: 'purchase'
      * @param amount [integer]: IssuingPurchase value in cents. Minimum = 0. ex: 1234 (= R$ 12.34)
      * @param tax [integer]: IOF amount taxed for international purchases. ex: 1234 (= R$ 12.34)
      * @param issuerAmount [integer]: issuer amount. ex: 1234 (= R$ 12.34)
@@ -27,45 +27,50 @@ class IssuingPurchase extends Resource {
      * @param merchantCurrencyCode [string]: merchant currency code. ex: 'USD'
      * @param merchantCurrencySymbol [string]: merchant currency symbol. ex: '$'
      * @param merchantCategoryCode [string]: merchant category code. ex: 'eatingPlacesRestaurants'
+     * @param merchantCategoryType [string]: merchant category type. ex: 'food'
      * @param merchantCountryCode [string]: merchant country code. ex: 'USA'
      * @param acquirerId [string]: acquirer ID. ex: '5656565656565656'
      * @param merchantId [string]: merchant ID. ex: '5656565656565656'
      * @param merchantName [string]: merchant name. ex: 'Google Cloud Platform'
-     * @param metadata [dictionary]: dictionary object used to store additional information about the IssuingPurchase object. ex: { authorizationId: 'OjZAqj' }
      * @param merchantFee [integer]: fee charged by the merchant to cover specific costs, such as ATM withdrawal logistics, etc. ex: 200 (= R$ 2.00)
      * @param walletId [string]: virtual wallet ID. ex: '5656565656565656'
      * @param methodCode [string]: method code. Options: 'chip', 'token', 'server', 'manual', 'magstripe' or 'contactless'
      * @param score [float]: internal score calculated for the authenticity of the purchase. ex: 7.6
      * @param endToEndId [string]: Unique id used to identify the transaction through all of its life cycle, even before the purchase is denied or accepted and gets its usual id. ex: '679cd385-642b-49d0-96b7-89491e1249a5'
      * @param tags [list of string]: list of strings for tagging. ex: ['travel', 'food']
-     * @param zipCode [string]: zip code of the merchant location. ex: "02101234"
      * 
      * Attributes (IssuingPurchase only):
+     * @param id [string]: unique id returned when IssuingPurchase is created. ex: '5656565656565656'
      * @param issuingTransactionIds [string]: ledger transaction ids linked to this Purchase
      * @param status [string]: current IssuingCard status. Options: 'approved', 'canceled', 'denied', 'confirmed' or 'voided'
+     * @param description [string]: IssuingPurchase description. ex: 'Office Supplies'
+     * @param metadata [dictionary object]: dictionary object used to store additional information about the IssuingPurchase object. ex: { authorizationId: 'OjZAqj' }.
+     * @param zipCode [string]: zip code of the merchant location. ex: '02101234'
      * @param created [string]: creation datetime for the IssuingPurchase. ex: '2020-03-10 10:30:00.000'
      * @param updated [string]: latest update datetime for the IssuingPurchase. ex: '2020-03-10 10:30:00.000'
      * 
      * Attributes (Authorization request only):
      * @param isPartialAllowed [bool]: true if the the merchant allows partial purchases. ex: false
-     * @param cardTags [list of strings]: tags of the IssuingCard responsible for this purchase. ex: ["travel", "food"]
-     * @param holderTags [list of strings]: tags of the IssuingHolder responsible for this purchase. ex: ["technology", "john snow"]
+     * @param cardTags [list of strings]: tags of the IssuingCard responsible for this purchase. ex: ['travel', 'food']
+     * @param holderId [string]: card holder ID. ex '5656565656565656' 
+     * @param holderTags [list of strings]: tags of the IssuingHolder responsible for this purchase. ex: ['technology', 'john snow']
      *
      */
     constructor({ 
-                    id=null, holderName=null, cardId=null, cardEnding=null, purpose=null, 
+                    holderName=null, productId=null, cardId=null, cardEnding=null, purpose=null, 
                     amount=null, tax=null, issuerAmount=null, issuerCurrencyCode=null, 
                     issuerCurrencySymbol=null, merchantAmount=null, merchantCurrencyCode=null, 
-                    merchantCurrencySymbol=null, merchantCategoryCode=null, merchantCountryCode=null, 
-                    acquirerId=null, merchantId=null, merchantName=null, metadata=null, merchantFee=null, 
-                    walletId=null, methodCode=null, score=null, endToEndId=null, 
-                    tags=null, zipCode=null, issuingTransactionIds=null, status=null, 
-                    created=null, updated=null, isPartialAllowed=null, cardTags=null, 
-                    holderTags=null 
+                    merchantCurrencySymbol=null, merchantCategoryCode=null, merchantCategoryType=null, 
+                    merchantCountryCode=null, acquirerId=null, merchantId=null, merchantName=null, 
+                    merchantFee=null, walletId=null, methodCode=null, score=null, endToEndId=null, 
+                    tags=null, id=null, issuingTransactionIds=null, status=null, description=null, 
+                    metadata=null, zipCode=null, created=null, updated=null, isPartialAllowed=null, 
+                    cardTags=null, holderId=null, holderTags=null 
                 }) {
         super(id);
 
         this.holderName = holderName;
+        this.productId = productId;
         this.cardId = cardId;
         this.cardEnding = cardEnding;
         this.purpose = purpose;
@@ -78,24 +83,27 @@ class IssuingPurchase extends Resource {
         this.merchantCurrencyCode = merchantCurrencyCode;
         this.merchantCurrencySymbol = merchantCurrencySymbol;
         this.merchantCategoryCode = merchantCategoryCode;
+        this.merchantCategoryType = merchantCategoryType;
         this.merchantCountryCode = merchantCountryCode;
         this.acquirerId = acquirerId;
         this.merchantId = merchantId;
         this.merchantName = merchantName;
-        this.metadata = metadata;
         this.merchantFee = merchantFee;
         this.walletId = walletId;
         this.methodCode = methodCode;
         this.score = score;
         this.endToEndId = endToEndId;
         this.tags = tags;
-        this.zipCode = zipCode;
         this.issuingTransactionIds = issuingTransactionIds;
         this.status = status;
-        this.updated = check.datetime(updated);
+        this.description = description;
+        this.metadata = metadata;
+        this.zipCode = zipCode;
         this.created = check.datetime(created);
+        this.updated = check.datetime(updated);
         this.isPartialAllowed = isPartialAllowed;
         this.cardTags = cardTags;
+        this.holderId = holderId;
         this.holderTags = holderTags;
     }
 }